change unix port to work on phoenix

+++ lib/libm_dbl/asinh.c
@@ -1,4 +1,4 @@
-#include "libm.h"
+#include "lib/libm_dbl/libm.h"
 
 /* asinh(x) = sign(x)*log(|x|+sqrt(x*x+1)) ~= x - x^3/6 + o(x^5) */
 double asinh(double x)
@@ -22,7 +22,7 @@ double asinh(double x)
 		x = log1p(x + x*x/(sqrt(x*x+1)+1));
 	} else {
 		/* |x| < 0x1p-26, raise inexact if x != 0 */
-		FORCE_EVAL(x + 0x1p120f);
+		FORCE_EVAL(x + (double)0x1p120f);
 	}
 	return s ? -x : x;
 }
